Bot_training/data_engineer/requirements.txt
psycopg2
pandas
matplotlib
sqlalchemy
python-dotenv
apache-airflow
requests
